requirements.txt
# ============================================================
# RAG知识库Agent系统依赖
# ============================================================

# LlamaIndex核心和扩展
llama-index>=0.10.0
llama-index-core>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-postprocessor-cohere-rerank>=0.1.0

# 向量存储
chromadb>=0.4.0

# PostgreSQL + pgvector
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
pgvector>=0.2.0
sqlalchemy>=2.0.0

# MinIO对象存储
minio>=7.2.0

# FastAPI后端
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.2.0

# 文档解析
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
markdown>=3.5.0
html2text>=2020.1.16
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=10.0.0

# 工具库
python-dotenv>=1.0.0
httpx>=0.26.0
tenacity>=8.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Redis缓存（可选）
redis>=5.0.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.23.0
